+++ Expanded Cheat Menu/Changelog.md
@@ -6,8 +6,9 @@
 ## v12.5 (Unreleased)
 ### Changed:
 - Map Exploration will no longer save the DeepScanAvailable setting.
-- Move Help>Interface/Screenshot to Game>.
+- Moved Help> (Interface and Screenshot) to Game>.
 - Log Errors In Console should have a better stack trace for threads.
+- Added a console input box to Console Log Window (and renamed it to Console Window).
 - Renamed Menu>Expanded CM to ECM:
 - > if you're missing some quickbar buttons, open up AppData\LocalStorage.lua and look for ToolbarItems.
 - Examine: